feat: store kill logs

+++ Isle.Domain/Entity/KillLog.cs
@@ -9,6 +9,6 @@ public class KillLog : BaseEntity<KillLog>, IPrefixedEntity
     public virtual Player? Killer { get; set; }
     public string? VictimId { get; set; }
     public virtual Player? Victim { get; set; }
-    public int VictimWeightKg { get; set; }
+    public double VictimWeightKg { get; set; }
     public static string Prefix { get; } = "Kill_log";
 }
